requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pinecone
sentence-transformers==2.2.2
openai==1.3.0
httpx==0.25.2
python-dotenv==1.0.0
PyMuPDF==1.23.8
pdfplumber==0.10.3
langchain-text-splitters==0.0.1
numpy==1.24.3
pandas==2.1.3
